+++ src/main/java/io/coodoo/workhorse/jobengine/control/event/AllJobsDoneEvent.java
@@ -2,6 +2,9 @@ package io.coodoo.workhorse.jobengine.control.event;
 
 import io.coodoo.workhorse.jobengine.entity.Job;
 
+/**
+ * @author coodoo GmbH (coodoo.io)
+ */
 public class AllJobsDoneEvent {
 
     private Job job;