+++ src/main/java/io/coodoo/workhorse/jobengine/entity/JobExecutionStatus.java
@@ -1,5 +1,8 @@
 package io.coodoo.workhorse.jobengine.entity;
 
+/**
+ * @author coodoo GmbH (coodoo.io)
+ */
 public enum JobExecutionStatus {
 
     QUEUED,